backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
langchain==0.0.335
langchain-groq==0.0.1
langchain-community==0.0.13
langchain-huggingface==0.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-multipart==0.0.6